+++ packages/aws-amplify-react/src/Auth/Provider/withOAuth.jsx
@@ -1,3 +1,16 @@
+/*
+ * Copyright 2017-2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance with
+ * the License. A copy of the License is located at
+ *
+ *     http://aws.amazon.com/apache2.0/
+ *
+ * or in the "license" file accompanying this file. This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+ * CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
+ * and limitations under the License.
+ */
+
 import * as React from 'react';
 import { Component } from 'react';
 
@@ -9,6 +22,7 @@ import {
     SignInButton, 
     SignInButtonContent
 } from '../../Amplify-UI/Amplify-UI-Components-React';
+import Constants from '../common/constants';
 
 const logger = new Logger('withOAuth');
 
@@ -24,10 +38,14 @@ export default function withOAuth(Comp, options) {
                 throw new Error('No Auth module found, please ensure @aws-amplify/auth is imported');
             }
 
-            const config = this.props.oauth_config || options || Auth.configure().oauth;
+            const { oauth={} } = Auth.configure();
+            // to keep backward compatibility
+            const cognitoHostedUIConfig = oauth? (oauth['domain']? oauth : oauth.awsCognito) : undefined;
+            const config = this.props.oauth_config || options || cognitoHostedUIConfig;
+
             logger.debug('withOAuth configuration', config);
             const { 
-                domain,  
+                domain, 
                 redirectSignIn,
                 redirectSignOut,
                 responseType
@@ -38,15 +56,21 @@ export default function withOAuth(Comp, options) {
                 + '/login?redirect_uri=' + redirectSignIn 
                 + '&response_type=' + responseType 
                 + '&client_id=' + (options.ClientId || Auth.configure().userPoolWebClientId);
-            window.location.assign(url);            
+
+            try {
+                localStorage.setItem(Constants.SIGN_IN_WITH_HOSTEDUI_KEY, 'true');
+            } catch (e) {
+                logger.debug('Failed to set item into localStorage', e);
+            }
+            window.location.assign(url);
         }
 
         render() {
             return (
                 <Comp {...this.props} OAuthSignIn={this.signIn} />
-            )
+            );
         }
-    }
+    };
 }
 
 const Button = (props) => (
@@ -60,6 +84,6 @@ const Button = (props) => (
             {I18n.get(props.label || 'Sign in with AWS')}
         </SignInButtonContent>
     </SignInButton>
-)
+);
 
 export const OAuthButton = withOAuth(Button);